change the long int by int

+++ printd_int.c
@@ -11,7 +11,7 @@ int printd_int(va_list args)
 	int length = 0;
 	int divisor = 1;
 
-	long int value = va_arg(args, long int);
+	int value = va_arg(args, int);
 
 	if (value == 0)
 	{
